requirements.txt
# RPLidar communication
rplidar-roboticia

# Serial port detection
pyserial

# 3D visualization and processing
open3d==0.19.0
numpy

# MQTT communication
paho-mqtt>=1.6.1

